Move the crypto stuff to a seperate file
